requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Cython==0.29.15
dlib==19.17.0
edgetpu==2.14.0
edgetpuvision==1.0
enum34==1.1.10
face-recognition==1.2.3
face-recognition-models==0.3.0
future==0.17.1
futures==3.1.1
gevent==1.5a1
greenlet==0.4.15
grpcio==1.27.2
grpcio-tools==1.27.2
httplib2==0.9.2
idna==2.8
msgpack-python==0.5.6
netifaces==0.10.4
numpy==1.18.1
pbr==5.2.0
Pillow==6.0.0
protobuf==3.11.3
pycurl==7.43.0
pygobject==3.22.0
PyJWT==1.4.2
pyparsing==2.4.6
pyserial==3.2.1
python-apt==1.4.1
python-debian==0.1.30
python-debianbts==2.6.1
pyzmq==18.0.1
reportbug==7.1.7
requests==2.22.0
scikit-learn==0.20.3
scipy==1.3.1
six==1.14.0
stevedore==1.30.1
svgwrite==1.3.1
tflite-runtime==2.1.0.post1
unattended-upgrades==0.1
urllib3==1.25.3
virtualenv==16.5.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
vitalsd==1.0
xgboost==0.90
zerorpc==0.6.1
